regroup gse alerts
