+++ gse-util/src/main/java/com/powsybl/gse/util/NodeChooser.java
@@ -396,14 +396,7 @@ public class NodeChooser<N, F extends N, D extends N, T extends N> extends GridP
         }
     }
 
-    private Alert nameAlreadyExistsAlert() {
-        Alert alert = new Alert(Alert.AlertType.ERROR);
-        alert.setTitle(RESOURCE_BUNDLE.getString("DragError"));
-        alert.setHeaderText(RESOURCE_BUNDLE.getString("Error"));
-        alert.setContentText(RESOURCE_BUNDLE.getString("DragFileExists"));
-        alert.showAndWait();
-        return alert;
-    }
+
 
     private void onMouseClickedEvent(MouseEvent event) {
         TreeItem<N> item = tree.getSelectionModel().getSelectedItem();
@@ -526,7 +519,7 @@ public class NodeChooser<N, F extends N, D extends N, T extends N> extends GridP
     private void accepTransferDrag(Folder folder, boolean s) {
         success = s;
         if (getCounter() >= 1) {
-            nameAlreadyExistsAlert();
+            GseAlerts.draggingError();
         } else if (getCounter() < 1) {
             Project monfichier = (Project) moveContext.source;
             monfichier.moveTo(folder);
@@ -633,23 +626,7 @@ public class NodeChooser<N, F extends N, D extends N, T extends N> extends GridP
     }
 
     public void createDeleteAlert(List<? extends TreeItem<N>> selectedTreeItems) {
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle(RESOURCE_BUNDLE.getString("ConfirmationDialog"));
-        String headerText;
-        if (selectedTreeItems.size() == 1) {
-            Node node = (Node) selectedTreeItems.get(0).getValue();
-            headerText = String.format(RESOURCE_BUNDLE.getString("FileWillBeDeleted"), node.getName());
-        } else if (selectedTreeItems.size() > 1) {
-            String names = selectedTreeItems.stream()
-                    .map(selectedItem -> selectedItem.getValue().toString())
-                    .collect(Collectors.joining(", "));
-            headerText = String.format(RESOURCE_BUNDLE.getString("FilesWillBeDeleted"), names);
-        } else {
-            throw new AssertionError();
-        }
-        alert.setHeaderText(headerText);
-        alert.setContentText(RESOURCE_BUNDLE.getString("DoYouConfirm"));
-        alert.showAndWait().ifPresent(result -> {
+        GseAlerts.deleteNodesAlert(selectedTreeItems).showAndWait().ifPresent(result -> {
             if (result == ButtonType.OK) {
                 setOnOkButton(selectedTreeItems);
             }
