+++ gse-util/src/main/resources/lang/GseAlerts.properties
@@ -1,0 +1,9 @@
+DragError=Drag error
+FileExists=The file name already exists
+Error=Error
+ConfirmationDialog=Confirmation Dialog
+FileWillBeDeleted=File %s will be deleted
+FilesWillBeDeleted=Files %s will be deleted
+DoYouConfirm=Do you confirm?
+
+
