+++ gse-util/src/main/resources/lang/GseAlerts_fr.properties
@@ -1,0 +1,6 @@
+DragError=Erreur de déplacement
+FileExists=Le nom du fichier existe déja
+Error=Erreur
+ConfirmationDialog=Fenêtre de confirmation
+FileWillBeDeleted=Le fichier %s sera supprimé
+FilesWillBeDeleted=Les fichiers %s seront supprimé
