+++ gse-util/src/main/resources/lang/NodeChooser.properties
@@ -4,13 +4,8 @@ Delete=Delete
 DeleteError=Delete error
 DeletedFolderContainsProjects=The folder %s contains at least one project
 Description=Description
-DoYouConfirm=Do you confirm?
-DragError=Drag error
-DragFileExists=The dragged file already exists
 Error=Error
 File=File
-FileWillBeDeleted=File %s will be deleted
-FilesWillBeDeleted=Files %s will be deleted
 Name=Name
 NewName=New Name
 OpenFile=Open file
