+++ gse-util/src/main/resources/lang/NodeChooser_fr.properties
@@ -4,13 +4,8 @@ Delete=Supprimer
 DeleteError=Erreur de Suppression
 DeletedFolderContainsProjects=le Dossier %s Contient au moins un projet
 Description=Description
-DoYouConfirm=Confirmez vous?
-DragError=Erreur de déplacement
-DragFileExists=le fichier déplacé existe déja
 Error=Erreur
 File=Fichier
-FileWillBeDeleted=Le fichier %s sera supprimé
-FilesWillBeDeleted=Les fichiers %s seront supprimés
 Name=Nom
 NewName=Entrez un nouveau nom
 OpenFile=Ouvrir un fichier
